Prevent fatal error in autoloader
Switches from require_once to include_once. The former throws a fatal error when a file isn't found while the later just emits a warning. Because of the fatal error autoloaders registered after the one provided by OMV won't be executed. The error message is also not quite obvious because of this, since it tells you which file that failed to load instead of which class.