Mantis 0001540: Error in datamodel conf.service.tftp
Note, we need to do a stupid workaround here when no users are defined. The problem is caused by the way how the configuration data is represented in XML and expected in JSON.

XML: <users></users>
JSON: { users: { user: [] } }

The workaround takes care that the internal representation will be fixed in the special case when no users are defined.