Try to fix http://forum.openmediavault.org/index.php/Thread/13897-cannot-open-dev-sr0-No-medium-found/?postID=125959#post125959.
Note, there are problems with USB network/modem dongles that have CDROM support for driver installation. The CDROM part is identified, and the above code is TRUE which indicates an inserted media, but blkid will fail with an error accessing the device. To prevent this we need to filter such devices. A real CDROM/DVD/... should have the ID_CDROM_MRW identifier.