The properties readOnly and editable can't be set via fieldDefaults() because many fields like textfield derive the value from a parent class. Thus we need to use defaults instead of fieldDefaults.
See https://www.sencha.com/forum/showthread.php?332560-Missing-readonly-cls-on-field-when-readOnly-is-set-in-fieldDefaults, https://www.sencha.com/forum/showthread.php?299104-FORM-with-readOnly-property-in-fieldDefaults-doesn-t-work.

The strange thing is that the OMV UI can reproduce this issue, either the ExtJS forum tolds the bug is fixed and i can not reproduce this behaviour with https://fiddle.sencha.com/#fiddle/bi5&view/editor.

Signed-off-by: Volker Theile <votdev@gmx.de>