Fixed ticket 1772: SMART overview show RED led, but all details are GREEN
When OpenMediaVault gets a drive's overall SMART status, it checks the normal
attributes, but then also checks the Reallocated_Sector_Ct (5) and
Current_Pending_Sector (197) attributes.

Then when viewing a drive's detailed SMART attributes, these same checks are
not performed. If you have a drive with at least 1 reallocated or pending
sector, this will cause the drive's overall status to be red, but all the
detailed SMART attributes will be green or grey.

To make the status consistent, I added checks to the prefailure attribute
checks, which will now cause those 2 attributes to show up as red when their
raw values are > 0.