Display Amiga Guru Meditation on HTTP 404 error page.
In contrast to the displayed text 'Press left mouse button to continue' the LMB is click is disabled on this page because it does not make any sense to hide the Guru Meditation. Changing the text is also no option because the Guru Meditation is a tribute to the Amiga Computer, thus it should look as much as possible as the original.

Signed-off-by: Volker Theile <votdev@gmx.de>
(cherry picked from commit 00562535722f7150dbb0ec03e790e2c71851bdd4)