Extract upstream version for comparison
distutils.version.LooseVersion fails while comparing versions like '5.0.0' < '5.0-1'. The result is a 'TypeError: unorderable types: int() < str()'. This is because the comparison of 0 and '-' does not work. Because of that the package revision must be stripped before the version strings are compared.

Fixes: https://github.com/openmediavault/openmediavault/issues/297

(cherry picked from commit be9768b23f737c28fc993688cf5cda2b7f3ef113)
Signed-off-by: Volker Theile <votdev@gmx.de>