Allow chrony to serve clients
This change provides a way for users to specify 'allow' directives to chrony so that OMV can serve time requests. When "Use NTP server" is selected, "Allowed clients" becomes available. This is an IP CIDR or host name list similar "Time servers." IP is better because host names must be resolvable when chrony starts. See chrony docs for more information.

Fixes: #730

Signed-off-by: Gregory Moyer <moyerg@syphr.com>