UDEV rules to workaround invalid serial numbers causes problems
The UDEV rules to workaround a JMicron issue with USB enclosures where serial ids duplicated does not work as expected. According to systemd/systemd#18111 the rules must be executed before 60-persistent-storage.rules to do not cause problems with filesystems.

The fix is to adapt the existing rules by removing the SYMLINK="..." part and rename it to 59-xxx.rules to be executed before 60-persistent-storage.rules. The openmediavault UDEV rules will then modify the ID_SERIAL and ID_SERIAL_SHORT properties of the affected devices and 60-persistent-storage.rules will take care about creating the device files.

This fix will affect some users that are using JMicron based enclosures or adaptors (e.g. Orangepi PC2) and that have applied the old openmediavault UDEV rules in the past. Such installations will break and need to be reconfigured from scratch related to filesystems and shared folders. The problem is that the filesystem configuration contains invalid device files after the new UDEV rules are applied.

Fixes: https://github.com/openmediavault/openmediavault/issues/906

Signed-off-by: Volker Theile <votdev@gmx.de>