Issue #921: Prevent using WLAN NICs for bond and bridge interfaces.
There is no way to configure the WLAN credentials when a WiFi interface is used as slave in a bond or bridge interface. Because of that WiFi interfaces are not listed in the UI as candidates for those interface types.
To get this implemented the UI and the backend must be completely refactored. If this is a required feature, then it's up to the community to contribute it in OMV6.

Fixes: https://github.com/openmediavault/openmediavault/issues/921

Signed-off-by: Volker Theile <votdev@gmx.de>
(cherry picked from commit bbba10dd1b926973745a8ee122afaa50dc15c2be)