Update to document faulty MB174U3S-4SB enclosure
Thanks for the effort in figuring out the workaround on this, I was able to use this stuff for my Linux Mint desktop. I came across the serial number handling for /dev/disk/by-id bug in an Icy Dock Black Vortex MB174U3S-4SB Quad Bay usb 3.0 eSATA external 3.5" SATA HDD Enclosure in the context of managing a ZFS pool including a few drives in this enclosure.

Although I am not using the Openmediavault project, I figure a reference to another affected hardware might leave a couple more breadcrumbs for people in the future running this problem down.

Signed-off-by: kr4z33 <kr4z33@yahoo.com>