Add 'autocapitalize' directive
I don't know if it works that way because i can not test it using a virtual keyboard or on a mobile device; but it makes the unit tests happy again.

The error that occurs without this directive:
NG0303: Can't bind to 'autocapitalize' since it isn't a known property of 'input'.

Signed-off-by: Volker Theile <votdev@gmx.de>