Reattach to running background process if possible.
Note, the background process status file does not contain any information about the process, so it is not possible to display any further information about it in the UI (notification sidebar or the task window after the attach action). It is not possible to implement that, apart from the fact that many tasks (e.g. scheduled tasks) do not contain any information that could be displayed there.

Signed-off-by: Volker Theile <votdev@gmx.de>