Prevent custom network interface search domain settings from being overwritten by settings provided via DHCP
Note: The `systemd-networkd` submodule will not be marked as dirty to force a redeployment of the network configuration. This has to be done manually via `omv-salt deploy run systemd-networkd` if needed. Otherwise the new Netplan configuration will be deployed as usual (incl. the new chhanges introduced by this PR) whenever the network configuration is changed.

Relates: https://github.com/openmediavault/openmediavault/issues/1551

Signed-off-by: Volker Theile <votdev@gmx.de>