Fix the max resize with of the several pages
Limit the width of the page content to max. 1280px. This is done to minimize scrolling with the mouse. The page content is aligned with reason on the left and not centered as is the case with other UIs (e.g. Home Assistant). If the content were centered, this would not really reduce the number of mouse movements, since the navigation bar is on the left side. Datatable and textarea pages will not be limited because they sometimes need more space, e.g. if many columns are displayed or the displayed text is not limited to 80 or any other number of chars.

The following page types are affected:

- Form
- RRD
- Selection list

Signed-off-by: Volker Theile <votdev@gmx.de>