Enhance database query to prevent an error when trying to run a SMART scheduled task via UI
There are situations that a device setting is stored in the database with a different devicefile. Running a scheduled task via UI does not take ALL devicefiles of a device into action which might cause a situation that the setting is not found.

Signed-off-by: Volker Theile <votdev@gmx.de>
(cherry picked from commit b0cc6a3c8adbc4df92e963bc739b4a9018f31756)
Signed-off-by: Volker Theile <votdev@gmx.de>