Add single node Kubernetes plugin (#1722)
Kubernetes is an open-source platform for orchestrating and managing containerized applications in an automated, scalable, and resource-efficient environment.

The plugin is intended to implement a single-node Kubernetes environment using K3s; nothing more. The LoadBalancer service is listening on port 8080 and 8443 by default.

The Kubernetes Dashboard is running on port 4443 by default. It is not possible to run it at https://foo:8443/kubernetes-dashboard for example because the Dashboard has problems with running from subpaths. At least i did not get it running with prefix stripping in Traefik.

Kubernetes was choosen because ...

- managing containers via podman and systemd is hell and complicated
- i don't want to be responsible for starting and stopping the containers
- there is already a Docker based plugin
- allow users to use Kubernetes workflows in a small scale environment
- make use of all benefits of K8s
- SUSE k3s is production ready for single node/edge scenarious

Automatic features:
- Create PVs for shared folders. Note, the PVs will not be removed when the shared folder configuration is removed in the UI.

Signed-off-by: Volker Theile <votdev@gmx.de>