Issue #1751: Handle empty objects correctly
Note, the whole handling of the `params` argument of the RPC implementation is faulty, but it can not be fixed without breaking existing code and behavior in core and all existing plugins. Therefore, the current implementation will be retained, even if it is wrong. The error is the processing of the `params` argument. This is not converted and processed as PHP objects but converted into an associative array. As a result, the information is lost as to whether it is an empty object or an empty array.

Relates to: https://github.com/openmediavault/openmediavault/issues/1751

Signed-off-by: Volker Theile <votdev@gmx.de>