Refactor RPC parameter handling
Support POST and GET for RPC calls. Please note that all RPCs can be called via GET and POST. However, it must be noted that there is a limit to the URL length for GET requests, so if a lot of data is to be transmitted, POST should be preferred.

Signed-off-by: Volker Theile <votdev@gmx.de>