Prevent unattended-upgrades from installing ALL packages
According to the unattended-upgrades README it should be possible to override the default shipped configuration. This is done by OMV, but it seems there is a bug in unattended-upgrades which appends instead of replaces the custom `Unattended-Upgrade::Origins-Pattern` configuration.

To workaround this issue, a `DPkg::Post-Invoke` is deleting the `/etc/apt/apt.conf.d/50unattended-upgrades` file.

Relates to: https://github.com/mvo5/unattended-upgrades/issues/366

Signed-off-by: Volker Theile <votdev@gmx.de>