Umount a file system via mountpoint
The mount point always exists in the response of the FileSystemMgmt::getListBg RPC which is used to populate the list of configured file systems in the UI. If a file system does not exist anymore, the mount point attribute still exists in the response data because it is fetched from the database. Because of that it makes sense to use this to unmount file systems from within the UI. Otherwise users can not get rid of those zombies.

Signed-off-by: Volker Theile <votdev@gmx.de>