feature: Move PageContext to injected service for all components
At the moment, the PageContext is being passed along all components. This is prone to errors, as it's hard to keep track of the necessary forwarding chain. Instead, the PageContext is now being injected by the Angular DI into all components.

Important: The PageContextService is declared as provider on each component that explicitly or implicitly use it (`providers: [PageContextService]`). This is important as the `ActivatedRoute` isn't available/correct outside the component context (e.g. when using the PageContextService as `providedIn: 'root'`).

Signed-off-by: David Dreschner <github-2017@dreschner.net>