Set filebrowser image tag to `v2.32.0`
It is not possible to use a newer version because generating a random password for the admin user has been introduced with v2.32.1 which is printed to stdout (visible via container logs). But these logs are not visible in OMV, so there is no way to tell the user which password to use. Additionally there is no way to bypass a custom password via env variable or command line arg.
:-(

Signed-off-by: Volker Theile <votdev@gmx.de>