Using smartctl to get hdd temperature values relied on the related smart IDs. Usually the IDs 190, 194 and 231 where used for temperature values. Some SandForce or Phison controlled SSDs, like the Corsair Force LS SSD, don't have any temperature sensors at all. This can lead to finding a wrong attribute at ID 231, like "SSD_LIFE_LEFT", giving the maintainer slight heart attacks.
getTemperature() used to solely check for one of the three IDs.
As, as far as my researches were correct, the attribute_name is always a string containing the substring "Temperature". Including a check for this substring should do the trick.
Changed the switch case to in_array, added a case insensitive check for substring "temperature", moved the comment blocks insided the switch cases to just before the loop.

Signed-off-by: robNice <robnice@downlord.net>